Backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
PyMuPDF>=1.23.8
chromadb>=0.4.15
google-generativeai>=0.3.2
sentence-transformers>=2.2.2
numpy>=1.24.3
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytest>=7.0.0
httpx>=0.24.0
aiofiles>=23.2.1
gtts>=2.5.4

# BERT QA dependencies
transformers>=4.35.0
torch>=2.0.0
